+++ crates/katana-core/tests/starknet.rs
@@ -22,6 +22,7 @@ fn create_test_starknet() -> StarknetWrapper {
     StarknetWrapper::new(StarknetConfig {
         seed: [0u8; 32],
         total_accounts: 2,
+        block_on_demand: false,
         allow_zero_max_fee: true,
         account_path: Some(test_account_path),
     })
