--- crates/katana-rpc/src/api/mod.rs
@@ -1,2 +1,0 @@
-pub mod katana;
-pub mod starknet;
