+++ crates/katana-rpc/src/config.rs
@@ -1,3 +1,4 @@
+#[derive(Debug, Clone)]
 pub struct RpcConfig {
     pub port: u16,
 }
