+++ crates/katana-rpc/src/utils/mod.rs
@@ -1,3 +1,5 @@
+#![allow(unused)]
+
 pub mod contract;
 pub mod state_diff;
 pub mod transaction;
